// resources/js/types/Models/Site.ts
export interface SiteInfo {
    texto_home: string | null;
    imagem_home: string | null;

    telefone: string | null;
    email: string | null;
    morada: string | null;
    slogan: string | null;
    sabado: string | null;
    heroText: string | null;
    horario_semana: string | null;
    horario_domingo: string | null;
}